absorb proximity into trajectories
