+++ ecoscope/base/__init__.py
@@ -5,7 +5,7 @@ from ecoscope.base._dataclasses import (
     RelocsSpeedFilter,
     TrajSegFilter,
 )
-from ecoscope.base.base import Relocations, Trajectory
+from ecoscope.base.base import Relocations, Trajectory, ProximityProfile, SpatialFeature
 from ecoscope.base.utils import (
     create_meshgrid,
     groupby_intervals,
@@ -21,4 +21,6 @@ __all__ = [
     "Trajectory",
     "create_meshgrid",
     "groupby_intervals",
+    "ProximityProfile",
+    "SpatialFeature",
 ]
