+++ tests/test_proiximity.py
@@ -1,5 +1,5 @@
 import pytest
-from ecoscope.analysis.proximity import SpatialFeature, Proximity, ProximityProfile
+from ecoscope.base import SpatialFeature, ProximityProfile
 
 
 @pytest.mark.skipif(not pytest.earthranger, reason="No connection to EarthRanger")
@@ -14,6 +14,6 @@ def test_proximity(er_io):
     relocations = er_io.get_subjectgroup_observations(subject_group_name=er_io.GROUP_NAME)
     trajectory = relocations.trajectories.from_relocations(relocations)
 
-    proximity_events = Proximity.calculate_proximity(proximity_profile=prox_profile, trajectory=trajectory)
+    proximity_events = trajectory.trajectories.calculate_proximity(prox_profile)
 
     assert len(proximity_events["spatialfeature_id"].unique()) == len(prox_profile.spatial_features)
